Fix insertion sort algorithm truncating 64 bit integers to 32 bit

+++ algorithms/insertion_sort/sort.c
@@ -2,9 +2,9 @@
 #include <stdint.h>
 
 void sort(int64_t *nums, size_t size) {
-    for (int i = 1; i < size; i++) {
-        int x = nums[i];
-        int j = i - 1;
+    for (size_t i = 1; i < size; i++) {
+        int64_t x = nums[i];
+        size_t j = i - 1;
 
         while (j >= 0 && nums[j] > x) {
             nums[j + 1] = nums[j];
